fix: workbench overview page display
