+++ src/types/wbSummary.ts
@@ -1,0 +1,10 @@
+// Helper types for generating workbench summary
+export type WorkbenchUpgradeSummaryItem = {
+	workbenchId: string;
+	targetTier: number;
+};
+
+export interface WorkbenchUpgradeSummary {
+	count: number;
+	usedIn: Array<WorkbenchUpgradeSummaryItem>;
+}
