+++ src/types/workbench.ts
@@ -12,14 +12,3 @@ export interface Workbench extends BaseItem {
 	baseTier: number;
 	tiers: WorkbenchTier[];
 }
-
-// Helper types for generating workbench summary
-export type WorkbenchUpgradeSummaryItem = {
-	workbenchId: string;
-	targetTier: number;
-};
-
-export interface WorkbenchUpgradeSummary {
-	count: number;
-	usedIn: Array<WorkbenchUpgradeSummaryItem>;
-}
